+++ .github/workflows/test.yml
@@ -1,0 +1,45 @@
+name: TEST
+
+on:
+  push:
+    branches: [ feat/add-hajun2 ]
+  workflow_dispatch:
+
+env:
+  APP_ENV: test
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      # Checkout the code
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      # Setup Python
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.10"
+
+      # Install Dependencies
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r scraper_server/requirements.txt
+
+      # Run Tests
+      - name: Test Code
+        env:
+          TESTING: true
+        run: |
+          pytest
+          zip -r9 report.zip htmlcov/
+
+      # Upload Test Results
+      - name: Upload Artifact
+        uses: actions/upload-artifact@v2
+        with:
+          name: REPORT
+          path: report.zip
